+++ Runtime/ui/painting/canvas_clip.cs
@@ -1,5 +1,6 @@
 using System.Collections.Generic;
 using Unity.UIWidgets.foundation;
+using Unity.UIWidgets.utils;
 using UnityEngine;
 
 namespace Unity.UIWidgets.ui {
