+++ Runtime/ui/painting/canvas_shader.cs
@@ -1,5 +1,6 @@
 using System;
 using Unity.UIWidgets.painting;
+using Unity.UIWidgets.utils;
 using UnityEngine;
 using UnityEngine.Rendering;
