+++ Runtime/ui/painting/path.cs
@@ -4,276 +4,10 @@ using System.Collections.Generic;
 //using System.Linq;
 using System.Text;
 using Unity.UIWidgets.foundation;
+using Unity.UIWidgets.utils;
 using UnityEngine;
-using UnityEngine.Analytics;
 
 namespace Unity.UIWidgets.ui {
-
-    public static class PathOptimizer {
-        public static bool optimizing = false;
-        public static int cmdNum = 0;
-    }
-
-    public interface Clearable {
-        void clear();
-    }
-    
-    
-    public class ClearableMaterialPropFlash {
-        static ClearableMaterialPropFlash _instance;
-
-        const int initial_size = 1024;
-        const int delta_size = 128;
-        int current_size = 0;
-
-        public static ClearableMaterialPropFlash instance {
-            get {
-                if (_instance != null) return _instance;
-                _instance = new ClearableMaterialPropFlash();
-                _instance.setup();
-                return _instance;
-            }
-        }
-
-        List<MaterialPropertyBlock> flash;
-        int curIndex;
-
-        void setup() {
-            this.flash = new List<MaterialPropertyBlock>(initial_size);
-            for (var i = 0; i < initial_size; i++) {
-                this.flash.Add(new MaterialPropertyBlock());
-            }
-
-            this.current_size = this.flash.Count;
-        }
-        
-        public MaterialPropertyBlock fetch() {
-            if (!PathOptimizer.optimizing) {
-                return new MaterialPropertyBlock();
-            }
-            
-            if (this.curIndex >= this.current_size) {
-                for (var i = 0; i < delta_size; i++) {
-                    this.flash.Add(new MaterialPropertyBlock());
-                }
-                this.current_size = this.flash.Count;
-            }
-
-            var ret = this.flash[this.curIndex++];
-            return ret;
-        }
-
-        public void clearAll() {
-            for (var i = 0; i < this.curIndex - 1; i++) {
-                this.flash[i].Clear();
-            }
-            
-            this.curIndex = 0;
-        }
-    }
-
-    public class ClearableSimpleFlash<T> where T : Clearable, new() {
-        static ClearableSimpleFlash<T> _instance;
-
-        const int initial_size = 1024;
-        const int delta_size = 128;
-        int current_size = 0;
-
-        public static ClearableSimpleFlash<T> instance {
-            get {
-                if (_instance != null) return _instance;
-                _instance = new ClearableSimpleFlash<T>();
-                _instance.setup();
-                return _instance;
-            }
-        }
-
-        List<T> flash;
-        int curIndex;
-
-        void setup() {
-            this.flash = new List<T>(initial_size);
-            for (var i = 0; i < initial_size; i++) {
-                this.flash.Add(new T());
-            }
-
-            this.current_size = this.flash.Count;
-        }
-        
-        public T fetch() {
-            if (!PathOptimizer.optimizing) {
-                return new T();
-            }
-            if (this.curIndex >= this.current_size) {
-                for (var i = 0; i < delta_size; i++) {
-                    this.flash.Add(new T());
-                }
-                this.current_size = this.flash.Count;
-            }
-
-            var ret = this.flash[this.curIndex++];
-            return ret;
-        }
-
-        public void clearAll() {
-            for (var i = 0; i < this.curIndex - 1; i++) {
-                this.flash[i].clear();
-            }
-            
-            this.curIndex = 0;
-        }
-    }
-
-    public class SimpleFlash<T> where T : new() {
-        static SimpleFlash<T> _instance;
-
-        const int initial_size = 1024;
-        const int delta_size = 128;
-        int current_size = 0;
-
-        public static SimpleFlash<T> instance {
-            get {
-                if (_instance != null) return _instance;
-                _instance = new SimpleFlash<T>();
-                _instance.setup();
-                return _instance;
-            }
-        }
-
-        List<T> flash;
-        int curIndex;
-
-        void setup() {
-            this.flash = new List<T>(initial_size);
-            for (var i = 0; i < initial_size; i++) {
-                this.flash.Add(new T());
-            }
-
-            this.current_size = this.flash.Count;
-        }
-        
-        public T fetch() {
-            if (!PathOptimizer.optimizing) {
-                return new T();
-            }
-            if (this.curIndex >= this.current_size) {
-                for (var i = 0; i < delta_size; i++) {
-                    this.flash.Add(new T());
-                }
-                this.current_size = this.flash.Count;
-            }
-
-            var ret = this.flash[this.curIndex++];
-            return ret;
-        }
-
-        public void clearAll() {
-            this.curIndex = 0;
-        }
-    }
-
-    public class Flash<T> {
-        static Flash<T> _instance;
-
-        const int initial_size = 1024;
-        const int delta_size = 128;
-        int current_size = 0;
-
-        public static Flash<T> instance {
-            get {
-                if (_instance != null) return _instance;
-                _instance = new Flash<T>();
-                _instance.setup();
-                return _instance;
-            }
-        }
-
-        List<List<T>> flash;
-        int curIndex;
-
-        void setup() {
-            this.flash = new List<List<T>>(initial_size);
-            for (var i = 0; i < initial_size; i++) {
-                this.flash.Add(new List<T>(256));
-            }
-
-            this.current_size = this.flash.Count;
-        }
-        
-        public List<T> fetch() {
-            if (!PathOptimizer.optimizing) {
-                return new List<T>();
-            }
-            
-            if (this.curIndex >= this.current_size) {
-                for (var i = 0; i < delta_size; i++) {
-                    this.flash.Add(new List<T>(256));
-                }
-
-                this.current_size = this.flash.Count;
-            }
-            
-
-            var ret = this.flash[this.curIndex++];
-            ret.Clear();
-            return ret;
-        }
-
-        public void clearAll() {
-            this.curIndex = 0;
-        }
-    }
-    
-    public class SimplePool<T> where T : new() {
-
-        static SimplePool<T> _instance;
-
-        const int initial_size = 256;
-        const int delta_size = initial_size / 2;
-        int current_size = 0;
-
-        public static SimplePool<T> instance {
-            get {
-                if (_instance != null) return _instance;
-                _instance = new SimplePool<T>();
-                _instance.setup();
-                return _instance;
-            }
-        }
-        
-        Stack<T> pool;
-
-        void setup() {
-            this.pool = new Stack<T>(initial_size * 2);
-            for (var i = 0; i < initial_size; i++) {
-                this.pool.Push(new T());
-            }
-
-            this.current_size = this.pool.Count;
-        }
-
-        public T fetch() {
-            if (this.current_size == 0) {
-                for (var i = 0; i < delta_size; i++) {
-                    this.pool.Push(new T());
-                }
-
-                this.current_size = this.pool.Count;
-            }
-
-            var ret = this.pool.Pop();
-            this.current_size--;
-            return ret;
-        }
-
-        public void recycle(T obj) {
-            if (obj != null) {
-                this.pool.Push(obj);
-                this.current_size++;
-            }
-        }
-    }
-    
     public class Path {
         const float _KAPPA90 = 0.5522847493f;
 
@@ -1343,7 +1077,7 @@ namespace Unity.UIWidgets.ui {
 
 
         public static PathPoint CreatePathPoint(float x, float y, float dx = 0, float dy = 0, float dmx = 0, float dmy = 0, PointFlags flags = PointFlags.corner) {
-            var ret = PathOptimizer.optimizing ? SimpleFlash<PathPoint>.instance.fetch() : new PathPoint();
+            var ret = GcCacheHelper.optimizing ? SimpleFlash<PathPoint>.instance.fetch() : new PathPoint();
             ret.x = x;
             ret.y = y;
             ret.dx = dx;
@@ -1416,7 +1150,7 @@ namespace Unity.UIWidgets.ui {
             this._distTol = 0.01f / scale;
             this._tessTol = 0.25f / scale;
 
-            if (PathOptimizer.optimizing) {
+            if (GcCacheHelper.optimizing) {
                 this._paths = Flash<PathPath>.instance.fetch();
                 this._points = Flash<PathPoint>.instance.fetch();
                 this._vertices = Flash<Vector3>.instance.fetch();
@@ -1605,7 +1339,7 @@ namespace Unity.UIWidgets.ui {
                 }
             }
 
-            var indices = PathOptimizer.optimizing ? Flash<int>.instance.fetch() : new List<int>(cindices);
+            var indices = GcCacheHelper.optimizing ? Flash<int>.instance.fetch() : new List<int>(cindices);
             for (var i = 0; i < this._paths.Count; i++) {
                 var path = this._paths[i];
                 if (path.count <= 2) {
